chore: refactor pack.get_tags -> codebase.revision.get_docker_tags
